make_ok

+++ srcs/color/color.c
@@ -6,7 +6,7 @@
 /*   By: nsakanou <nsakanou@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/03 17:03:05 by nsakanou          #+#    #+#             */
-/*   Updated: 2024/07/15 10:34:28 by nsakanou         ###   ########.fr       */
+/*   Updated: 2024/07/19 15:31:49 by nsakanou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,14 +19,14 @@ void	init_rgb(t_rgb *rgb, double r, double g, double b)
 	rgb->blue = b;
 }
 
-uint32_t	encode_color(t_rgb color)
+int	encode_color(t_rgb color)
 {
-	u_int8_t	red;
-	u_int8_t	green;
-	u_int8_t	blue;
+	int	red;
+	int	green;
+	int	blue;
 
-	red = (u_int8_t)(color.red * 255);
-	green = (u_int8_t)(color.green * 255);
-	blue = (u_int8_t)(color.blue * 255);
+	red = (int)(color.red * 255);
+	green = (int)(color.green * 255);
+	blue = (int)(color.blue * 255);
 	return ((red << 16) | (green << 8) | blue);
 }
